Fix ChefSpec
