+++ spec/vendor_install_spec.rb
@@ -1,6 +1,6 @@
 require_relative 'spec_helper'
 
-describe 'install_varnish::distro_install' do
+describe 'install_varnish::vendor_install' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new(step_into: %w(varnish_repo varnish_config vcl_template vcl_file varnish_log)) do |node|
       node_resources(node)
@@ -12,7 +12,7 @@ describe 'install_varnish::distro_install' do
   end
 
   it 'does not configure the varnish vendor repo' do
-    expect(chef_run).to_not configure_varnish_repo('vendor')
+    expect(chef_run).to configure_varnish_repo('configure')
   end
 
   it 'installs varnish' do
@@ -28,7 +28,6 @@ describe 'install_varnish::distro_install' do
 
   it 'creates the default varnish config and VCLs' do
     expect(chef_run).to configure_vcl_template('default.vcl')
-    expect(chef_run).to create_template('/etc/default/varnish')
     expect(chef_run).to create_template('/etc/varnish/default.vcl')
   end
 
@@ -38,8 +37,6 @@ describe 'install_varnish::distro_install' do
   end
 
   it 'creates the log & ncsa template, enables the varnishlog and varnishncsa service' do
-    expect(chef_run).to create_template('/etc/default/varnishlog')
-    expect(chef_run).to create_template('/etc/default/varnishncsa')
     expect(chef_run).to enable_service('varnishlog')
     expect(chef_run).to enable_service('varnishncsa')
   end
